Don't instrument if instrumented in previous pass
There are 2 passes that do nearly the same thing: the first pass checks
for the WatchDog argument while the second pass will check for a @Watch
annotation and use TLS storage to get the WatchDog object. If you have
both the argument and annotation, DO NOT INSTRUMENT TWICE. Only the
argument instrumentation pass will be applied, the annotation pass will
be ignored.

Remember that the argument pass will add code at the beginning of the
method to check null-check the arg. If it is null, the code will
attempt to load the WatchDog object from TLS.