Switched to loop instrumentation
Switched branch instrumentation to loop instrumentation. This generates
better performing insturmentation logic and works around potential bugs
in try-with-resource blocks.

Remember that try-with-resource blocks generate branches when closing
the resources. In the old branch instrumenter, it was possible to break
out while resources were being cleaned up. That's something we never
want. The loop instrumenter only instruments if the operation results in
a branch that could potentially loop.